refactoring, v2 disabled by default, aws region as setting
