+++ src/main/resources/application.conf
@@ -20,6 +20,7 @@ airlock {
             secretkey = ${?AIRLOCK_STORAGE_S3_ADMIN_SECRETKEY}
         }
         region = ${?AIRLOCK_STORAGE_S3_AWS_REGION}
+        v2SignatureEnabled = ${?AIRLOCK_STORAGE_S3_V2_ENABLED}
     }
 
     sts {
