+++ src/main/resources/reference.conf
@@ -28,6 +28,7 @@ airlock {
             secretkey = "secretkey"
         }
         region = "us-east-1"
+        v2SignatureEnabled = false
     }
 
     sts {
